MapChips内の各クラスを別ファイルに分割
