+++ packages/tiled-map/src/main.ts
@@ -1,6 +1,7 @@
 export { TiledMap, TiledMapProperties } from './TiledMap'
 export { TiledMapData, TiledMapDataItem, TiledMapDataProperties } from './MapData/TiledMapData'
-export { MapChipImage, MapChipImageProperties, MapChipsCollection, MapChipCollectionProperties } from './MapChips'
+export { MapChipImage, MapChipImageProperties } from './MapChipImage'
+export { MapChipsCollection, MapChipCollectionProperties } from './MapChipsCollection'
 export { MapChipFragment, MapChipFragmentProperties, MapChip, AutoTileMapChip, isAutoTileMapChip } from './MapChip'
 export { AutoTiles, AutoTileProperties, AutoTile, AutoTilesProperties } from './AutoTile/AutoTiles'
 export { DefaultAutoTileImportStrategy } from './AutoTile/DefaultAutoTileImportStrategy'
